Restructured algo No change added to the algo
